Move the files to a library
